+++ NEWS.md
@@ -56,6 +56,7 @@
 - Updated behavior of formatting functions `jjcsformat_count_denom_fraction` and `jjcsformat_fraction_count_denom` 
 and introduce functions `jjcsformat_count_denom_fraction_legacy` and `jjcsformat_fraction_count_denom_legacy` for backward compatibility reasons (#450)
 - Updated `insightsengineering` urls to `pharmaverse`
+- Added factor assertion on `.var` in `s_freq_j` (#453).
 
 ### Added
 - Added `strict_match()` for uniquely matching a value in the odd or even positions of a character vector.